# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - UF munícipios",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/municipios/{uf}": {
            "get": {
                "tags": [
                    "Municípios"
                ],
                "summary": "Obter todos os municípios por estado",
                "description": "Recupera uma lista de municípios no estado especificado (UF).",
                "operationId": "d5f1f29d88272f7a2d55b5140cae3d83",
                "parameters": [
                    {
                        "name": "uf",
                        "in": "path",
                        "description": "O UF (código do estado) para o qual recuperar os municípios.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar pelo nome do município.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filtrar por email (se aplicável).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de resultados por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uma lista de municípios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "nome": {
                                                        "type": "string"
                                                    },
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "ibge_code": {
                                                        "type": "integer",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida"
                    },
                    "500": {
                        "description": "Erro Interno do Servidor"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Municípios",
            "description": "Municípios"
        }
    ]
}